좋아요 부분 색상변경 기능 개선
